Revert "fix(android): Updated removeListeners argument type (#619)"
This reverts commit c01a23de3e7df9445b44f46f972c1d6c0caaffb4.

+++ android/src/main/java/com/reactnativecommunity/cameraroll/CameraRollModule.java
@@ -1098,5 +1098,5 @@ public class CameraRollModule extends NativeCameraRollModuleSpec {
   }
 
   public void addListener(String eventName) {}
-  public void removeListeners(int count) {}
+  public void removeListeners(double count) {}
 }
